moved to module as per sam recommendation
